添加创建文件系统调用sys_open，修复lib/string.c中的bug

+++ inc/fs.h
@@ -39,6 +39,6 @@ extern struct partition *cur_part;
 
 void filesys_init(void);
 int32_t path_depth_cnt(char *pathname);
-int32_t sys_open(const char *pathname, uint32_t flags);
+int32_t sys_open(const char *pathname, uint8_t flags);
 
 #endif
